@@ -1,22 +1,27 @@
+"""Reproducibility helpers for manifests and run bundles."""
+
 from __future__ import annotations
 
 import hashlib
 import json
-import os
-import pathlib
 import platform
+import shutil
 import sys
 import time
 from collections.abc import Mapping
-from typing import Any
+from contextlib import suppress
+from dataclasses import dataclass
+from importlib import metadata as importlib_metadata
+from importlib.metadata import PackageNotFoundError
+from pathlib import Path
 
 import pyarrow as pa
 
-try:
-    import importlib.metadata as importlib_metadata  # py3.8+
-except Exception:  # pragma: no cover
-    import importlib_metadata  # type: ignore
-
+from arrowdsl.contracts import Contract, DedupeSpec, SortKey
+from core_types import JsonDict, JsonValue, PathLike, ensure_path
+from relspec.compiler import CompiledOutput
+from relspec.model import RelationshipRule
+from relspec.registry import ContractCatalog, DatasetLocation, RelationshipRegistry
 
 # -----------------------
 # Basic environment capture
@@ -26,38 +31,41 @@ except Exception:  # pragma: no cover
 def _pkg_version(name: str) -> str | None:
     try:
         return importlib_metadata.version(name)
-    except Exception:
+    except PackageNotFoundError:
         return None
 
 
-def try_get_git_info(repo_root: str | None) -> dict[str, Any]:
-    """
-    Best-effort git info without shelling out.
+def try_get_git_info(repo_root: str | None) -> JsonDict:
+    """Collect best-effort git info without shelling out.
+
     Looks for:
       .git/HEAD and referenced ref file.
+
+    Returns
+    -------
+    JsonDict
+        Git metadata when available, otherwise a minimal status dict.
     """
     if not repo_root:
         return {"present": False}
 
-    git_dir = os.path.join(repo_root, ".git")
-    head_path = os.path.join(git_dir, "HEAD")
-    if not pathlib.Path(head_path).exists():
+    git_dir = Path(repo_root) / ".git"
+    head_path = git_dir / "HEAD"
+    if not head_path.exists():
         return {"present": False}
 
     try:
-        head = pathlib.Path(head_path).open("r", encoding="utf-8").read().strip()
-    except Exception:
+        head = head_path.read_text(encoding="utf-8").strip()
+    except OSError:
         return {"present": True, "error": "failed_to_read_HEAD"}
 
     if head.startswith("ref:"):
         ref = head.split(":", 1)[1].strip()
-        ref_path = os.path.join(git_dir, ref.replace("/", os.sep))
-        sha = None
-        if pathlib.Path(ref_path).exists():
-            try:
-                sha = pathlib.Path(ref_path).open("r", encoding="utf-8").read().strip()
-            except Exception:
-                sha = None
+        ref_path = git_dir / Path(ref)
+        sha: str | None = None
+        if ref_path.exists():
+            with suppress(OSError):
+                sha = ref_path.read_text(encoding="utf-8").strip()
         return {"present": True, "head": head, "ref": ref, "commit": sha}
 
     # Detached head case: HEAD contains commit sha
@@ -65,14 +73,18 @@ def try_get_git_info(repo_root: str | None) -> dict[str, Any]:
 
 
 def collect_repro_info(
-    repo_root: str | None = None, *, extra: dict[str, Any] | None = None
-) -> dict[str, Any]:
-    """
-    Captures a compact reproducibility bundle.
+    repo_root: str | None = None, *, extra: Mapping[str, JsonValue] | None = None
+) -> JsonDict:
+    """Capture a compact reproducibility bundle.
 
     Intentionally avoids huge payloads (pip freeze, full env vars).
+
+    Returns
+    -------
+    JsonDict
+        Reproducibility metadata snapshot.
     """
-    info: dict[str, Any] = {
+    info: JsonDict = {
         "python": {
             "version": sys.version,
             "executable": sys.executable,
@@ -99,11 +111,11 @@ def collect_repro_info(
 # -----------------------
 
 
-def _ensure_dir(path: str) -> None:
-    pathlib.Path(path).mkdir(exist_ok=True, parents=True)
+def _ensure_dir(path: Path) -> None:
+    path.mkdir(exist_ok=True, parents=True)
 
 
-def _json_default(obj: Any) -> Any:
+def _json_default(obj: object) -> JsonValue:
     # Arrow types
     if isinstance(obj, pa.Schema):
         return arrow_schema_to_dict(obj)
@@ -112,28 +124,26 @@ def _json_default(obj: Any) -> Any:
 
     # Common containers
     if isinstance(obj, set):
-        return sorted(list(obj))
+        return sorted((_json_default(v) for v in obj), key=str)
     if isinstance(obj, tuple):
-        return list(obj)
+        return [_json_default(v) for v in obj]
 
     # Dataclasses / objects
-    if hasattr(obj, "__dict__"):
-        # best-effort shallow dict
-        try:
-            return {k: _json_default(v) for k, v in obj.__dict__.items()}
-        except Exception:
-            return str(obj)
+    obj_dict = getattr(obj, "__dict__", None)
+    if isinstance(obj_dict, Mapping):
+        return {str(k): _json_default(v) for k, v in obj_dict.items()}
 
     return str(obj)
 
 
-def _write_json(path: str, data: Any, *, overwrite: bool = True) -> str:
-    _ensure_dir(os.path.dirname(path) or ".")
-    if overwrite and pathlib.Path(path).exists():
-        pathlib.Path(path).unlink()
-    with pathlib.Path(path).open("w", encoding="utf-8") as f:
+def _write_json(path: PathLike, data: JsonValue, *, overwrite: bool = True) -> str:
+    target = ensure_path(path)
+    _ensure_dir(target.parent)
+    if overwrite and target.exists():
+        target.unlink()
+    with target.open("w", encoding="utf-8") as f:
         json.dump(data, f, indent=2, ensure_ascii=False, sort_keys=True, default=_json_default)
-    return path
+    return str(target)
 
 
 # -----------------------
@@ -141,7 +151,14 @@ def _write_json(path: str, data: Any, *, overwrite: bool = True) -> str:
 # -----------------------
 
 
-def arrow_schema_to_dict(schema: pa.Schema) -> dict[str, Any]:
+def arrow_schema_to_dict(schema: pa.Schema) -> JsonDict:
+    """Serialize an Arrow schema to a plain dictionary.
+
+    Returns
+    -------
+    JsonDict
+        JSON-serializable schema representation.
+    """
     return {
         "fields": [
             {"name": f.name, "type": str(f.type), "nullable": bool(f.nullable)} for f in schema
@@ -150,176 +167,153 @@ def arrow_schema_to_dict(schema: pa.Schema) -> dict[str, Any]:
 
 
 def schema_fingerprint(schema: pa.Schema) -> str:
+    """Compute a stable schema fingerprint hash.
+
+    Returns
+    -------
+    str
+        SHA-256 fingerprint of the schema.
+    """
     payload = json.dumps(arrow_schema_to_dict(schema), sort_keys=True).encode("utf-8")
     return hashlib.sha256(payload).hexdigest()
 
 
-def serialize_contract(contract: Any) -> dict[str, Any]:
-    """
-    Serialize an arrowdsl.contracts.Contract-like object.
-    Works with our earlier Contract shape, but is intentionally defensive.
-    """
-    schema = getattr(contract, "schema", None)
-    sch_dict = arrow_schema_to_dict(schema) if isinstance(schema, pa.Schema) else None
-    sch_fp = schema_fingerprint(schema) if isinstance(schema, pa.Schema) else None
+def serialize_contract(contract: Contract) -> JsonDict:
+    """Serialize a ``Contract`` into JSON-friendly metadata.
 
-    dedupe = getattr(contract, "dedupe", None)
-    canonical_sort = getattr(contract, "canonical_sort", None)
+    Returns
+    -------
+    JsonDict
+        Serialized contract metadata.
+    """
 
-    def _serialize_sort_keys(keys: Any) -> Any:
+    def _serialize_sort_keys(keys: tuple[SortKey, ...] | None) -> list[JsonDict] | None:
         if keys is None:
             return None
-        out = []
-        for k in keys:
-            col = getattr(k, "column", None)
-            order = getattr(k, "order", None)
-            if col is None and isinstance(k, dict):
-                col = k.get("column")
-                order = k.get("order")
-            out.append({"column": col, "order": order})
-        return out
+        return [{"column": key.column, "order": key.order} for key in keys]
 
-    dedupe_dict = None
-    if dedupe is not None:
+    dedupe_dict: JsonDict | None = None
+    if contract.dedupe is not None:
+        dedupe: DedupeSpec = contract.dedupe
         dedupe_dict = {
-            "keys": list(getattr(dedupe, "keys", []) or []),
-            "strategy": getattr(dedupe, "strategy", None),
-            "tie_breakers": _serialize_sort_keys(getattr(dedupe, "tie_breakers", None)),
+            "keys": list(dedupe.keys),
+            "strategy": dedupe.strategy,
+            "tie_breakers": _serialize_sort_keys(dedupe.tie_breakers),
         }
 
     return {
-        "name": getattr(contract, "name", None),
-        "version": getattr(contract, "version", None),
-        "required_non_null": list(getattr(contract, "required_non_null", []) or []),
-        "key_fields": list(getattr(contract, "key_fields", []) or []),
-        "schema_fingerprint": sch_fp,
-        "schema": sch_dict,
+        "name": contract.name,
+        "version": contract.version,
+        "required_non_null": list(contract.required_non_null),
+        "key_fields": list(contract.key_fields),
+        "schema_fingerprint": schema_fingerprint(contract.schema),
+        "schema": arrow_schema_to_dict(contract.schema),
         "dedupe": dedupe_dict,
-        "canonical_sort": _serialize_sort_keys(canonical_sort),
+        "canonical_sort": _serialize_sort_keys(contract.canonical_sort),
     }
 
 
-def serialize_contract_catalog(contract_catalog: Any) -> dict[str, Any]:
+def serialize_contract_catalog(contract_catalog: ContractCatalog) -> JsonDict:
+    """Serialize a ``ContractCatalog`` snapshot.
+
+    Returns
+    -------
+    JsonDict
+        Serialized contract catalog.
     """
-    Serialize a ContractCatalog-like object from relspec/registry.py
+    names = list(contract_catalog.names())
+    return {"contracts": [serialize_contract(contract_catalog.get(name)) for name in names]}
+
+
+def serialize_relationship_rule(rule: RelationshipRule) -> JsonDict:
+    """Serialize a ``RelationshipRule`` into JSON-friendly metadata.
+
+    Returns
+    -------
+    JsonDict
+        Serialized rule metadata.
     """
-    names = []
-    if hasattr(contract_catalog, "names"):
-        names = list(contract_catalog.names())
-    elif hasattr(contract_catalog, "_contracts"):
-        names = sorted(list(contract_catalog._contracts.keys()))  # type: ignore
-
-    out = {"contracts": []}
-    for name in names:
-        c = (
-            contract_catalog.get(name)
-            if hasattr(contract_catalog, "get")
-            else contract_catalog._contracts[name]
-        )  # type: ignore
-        out["contracts"].append(serialize_contract(c))
-    return out
-
-
-def serialize_relationship_rule(rule: Any) -> dict[str, Any]:
-    """
-    Serialize a relspec.model.RelationshipRule-like object.
-    """
-
-    def _drefs() -> list[str]:
-        ins = getattr(rule, "inputs", None) or []
-        names = []
-        for d in ins:
-            nm = getattr(d, "name", None)
-            names.append(nm if nm is not None else str(d))
-        return names
-
+    inputs = [ref.name for ref in rule.inputs]
     return {
-        "name": getattr(rule, "name", None),
-        "kind": str(getattr(rule, "kind", None)),
-        "output_dataset": getattr(rule, "output_dataset", None),
-        "contract_name": getattr(rule, "contract_name", None),
-        "priority": int(getattr(rule, "priority", 0)),
-        "emit_rule_meta": bool(getattr(rule, "emit_rule_meta", True)),
-        "inputs": _drefs(),
-        # configs (best-effort)
-        "hash_join": getattr(
-            getattr(rule, "hash_join", None), "__dict__", getattr(rule, "hash_join", None)
-        ),
-        "interval_align": getattr(
-            getattr(rule, "interval_align", None), "__dict__", getattr(rule, "interval_align", None)
-        ),
-        "project": getattr(
-            getattr(rule, "project", None), "__dict__", getattr(rule, "project", None)
-        ),
-        "post_kernels": [
-            getattr(k, "__dict__", str(k)) for k in (getattr(rule, "post_kernels", None) or [])
-        ],
+        "name": rule.name,
+        "kind": str(rule.kind),
+        "output_dataset": rule.output_dataset,
+        "contract_name": rule.contract_name,
+        "priority": int(rule.priority),
+        "emit_rule_meta": bool(rule.emit_rule_meta),
+        "inputs": inputs,
+        "hash_join": _json_default(rule.hash_join) if rule.hash_join is not None else None,
+        "interval_align": _json_default(rule.interval_align)
+        if rule.interval_align is not None
+        else None,
+        "project": _json_default(rule.project) if rule.project is not None else None,
+        "post_kernels": [_json_default(k) for k in rule.post_kernels],
     }
 
 
-def serialize_relationship_registry(registry: Any) -> dict[str, Any]:
-    """
-    Serialize a RelationshipRegistry-like object from relspec/registry.py
-    """
-    rules = []
-    if hasattr(registry, "rules"):
-        rules = registry.rules()
-    elif hasattr(registry, "_rules_by_name"):
-        rules = list(registry._rules_by_name.values())  # type: ignore
+def serialize_relationship_registry(registry: RelationshipRegistry) -> JsonDict:
+    """Serialize a ``RelationshipRegistry`` snapshot.
 
+    Returns
+    -------
+    JsonDict
+        Serialized relationship registry.
+    """
     return {
-        "rules": [serialize_relationship_rule(r) for r in rules],
-        "outputs": list(registry.outputs()) if hasattr(registry, "outputs") else None,
-        "inputs": list(registry.inputs()) if hasattr(registry, "inputs") else None,
+        "rules": [serialize_relationship_rule(r) for r in registry.rules()],
+        "outputs": list(registry.outputs()),
+        "inputs": list(registry.inputs()),
     }
 
 
-def serialize_dataset_locations(locations: Mapping[str, Any]) -> dict[str, Any]:
+def serialize_dataset_locations(locations: Mapping[str, DatasetLocation]) -> JsonDict:
+    """Serialize dataset locations returned by persistence helpers.
+
+    Returns
+    -------
+    JsonDict
+        Serialized dataset location mapping.
     """
-    Serialize the mapping returned by persist_relspec_input_datasets.
-    Supports either DatasetLocation dataclass objects or plain dicts.
-    """
-    out: dict[str, Any] = {}
+    out: JsonDict = {}
     for name, loc in locations.items():
-        if isinstance(loc, dict):
-            out[name] = dict(loc)
-        else:
-            out[name] = {
-                "path": getattr(loc, "path", None),
-                "format": getattr(loc, "format", None),
-                "partitioning": getattr(loc, "partitioning", None),
-            }
+        out[name] = {
+            "path": str(loc.path),
+            "format": loc.format,
+            "partitioning": loc.partitioning,
+        }
     return out
 
 
-def serialize_compiled_outputs(compiled: Mapping[str, Any]) -> dict[str, Any]:
+def serialize_compiled_outputs(compiled: Mapping[str, CompiledOutput]) -> JsonDict:
+    """Serialize compiled relationship outputs.
+
+    Returns
+    -------
+    JsonDict
+        Serialized compiled outputs.
     """
-    Serialize compiled relationship outputs (CompiledOutput-like objects).
-    """
-    out: dict[str, Any] = {"outputs": {}}
+    outputs: dict[str, JsonDict] = {}
     for key, obj in compiled.items():
-        # CompiledOutput(output_dataset, contract_name, contributors)
-        rec = {
-            "output_dataset": getattr(obj, "output_dataset", key),
-            "contract_name": getattr(obj, "contract_name", None),
-            "contributors": [],
-        }
-        contributors = getattr(obj, "contributors", None) or []
-        for cr in contributors:
-            rule = getattr(cr, "rule", None)
-            rec["contributors"].append(
+        contributors: list[JsonDict] = []
+        for compiled_rule in obj.contributors:
+            rule = compiled_rule.rule
+            contributors.append(
                 {
-                    "rule_name": getattr(rule, "name", None),
-                    "priority": getattr(rule, "priority", None),
-                    "kind": str(getattr(rule, "kind", None)),
-                    "inputs": [
-                        getattr(d, "name", str(d)) for d in (getattr(rule, "inputs", None) or [])
-                    ],
+                    "rule_name": rule.name,
+                    "priority": rule.priority,
+                    "kind": str(rule.kind),
+                    "inputs": [ref.name for ref in rule.inputs],
                 }
             )
-        out["outputs"][key] = rec
-    out["produced_output_keys"] = sorted(list(compiled.keys()))
-    return out
+        outputs[key] = {
+            "output_dataset": obj.output_dataset,
+            "contract_name": obj.contract_name,
+            "contributors": contributors,
+        }
+    return {
+        "outputs": outputs,
+        "produced_output_keys": sorted(compiled.keys()),
+    }
 
 
 # -----------------------
@@ -327,11 +321,27 @@ def serialize_compiled_outputs(compiled: Mapping[str, Any]) -> dict[str, Any]:
 # -----------------------
 
 
-def make_run_bundle_name(run_manifest: Mapping[str, Any], run_config: Mapping[str, Any]) -> str:
+def make_run_bundle_name(
+    run_manifest: Mapping[str, JsonValue], run_config: Mapping[str, JsonValue]
+) -> str:
+    """Build a deterministic-ish run bundle name.
+
+    Returns
+    -------
+    str
+        Bundle name in the form run_<created_at>_<hash10>.
     """
-    Deterministic-ish bundle name: run_<created_at>_<hash10>
-    """
-    ts = int(run_manifest.get("created_at_unix_s") or time.time())
+    created_value = run_manifest.get("created_at_unix_s")
+    if isinstance(created_value, bool):
+        created_value = None
+    if isinstance(created_value, int):
+        ts = created_value
+    elif isinstance(created_value, float) or (
+        isinstance(created_value, str) and created_value.isdigit()
+    ):
+        ts = int(created_value)
+    else:
+        ts = int(time.time())
     payload = json.dumps(
         {"manifest": run_manifest, "config": run_config}, sort_keys=True, default=str
     ).encode("utf-8")
@@ -339,27 +349,152 @@ def make_run_bundle_name(run_manifest: Mapping[str, Any], run_config: Mapping[st
     return f"run_{ts}_{h}"
 
 
+@dataclass(frozen=True)
+class RunBundleContext:
+    """Inputs required to write a reproducible run bundle."""
+
+    base_dir: PathLike
+    run_manifest: Mapping[str, JsonValue]
+    run_config: Mapping[str, JsonValue]
+
+    relationship_registry: RelationshipRegistry | None = None
+    relationship_contracts: ContractCatalog | None = None
+    compiled_relationship_outputs: Mapping[str, CompiledOutput] | None = None
+
+    relspec_input_locations: Mapping[str, DatasetLocation] | None = None
+    relspec_input_tables: Mapping[str, pa.Table] | None = None
+    relationship_output_tables: Mapping[str, pa.Table] | None = None
+    cpg_output_tables: Mapping[str, pa.Table] | None = None
+
+    include_schemas: bool = True
+    overwrite: bool = True
+
+
+def _ensure_bundle_dir(bundle_dir: Path, *, overwrite: bool) -> None:
+    if overwrite and bundle_dir.exists():
+        with suppress(OSError):
+            shutil.rmtree(bundle_dir)
+    _ensure_dir(bundle_dir)
+
+
+def _write_manifest_files(
+    bundle_dir: Path,
+    context: RunBundleContext,
+    *,
+    bundle_name: str,
+    files_written: list[str],
+) -> None:
+    files_written.append(
+        _write_json(bundle_dir / "manifest.json", dict(context.run_manifest), overwrite=True)
+    )
+    files_written.append(
+        _write_json(bundle_dir / "config.json", dict(context.run_config), overwrite=True)
+    )
+    repo_root_value = context.run_config.get("repo_root")
+    repo_root = repo_root_value if isinstance(repo_root_value, str) else None
+    repro = collect_repro_info(repo_root=repo_root, extra={"bundle_name": bundle_name})
+    files_written.append(_write_json(bundle_dir / "repro.json", repro, overwrite=True))
+
+
+def _write_dataset_locations(
+    bundle_dir: Path, context: RunBundleContext, files_written: list[str]
+) -> None:
+    if context.relspec_input_locations is None:
+        return
+    ds_dir = bundle_dir / "datasets"
+    _ensure_dir(ds_dir)
+    locs = serialize_dataset_locations(context.relspec_input_locations)
+    files_written.append(_write_json(ds_dir / "locations.json", locs, overwrite=True))
+
+
+def _write_relspec_snapshots(
+    bundle_dir: Path, context: RunBundleContext, files_written: list[str]
+) -> None:
+    relspec_dir = bundle_dir / "relspec"
+    _ensure_dir(relspec_dir)
+    if context.relationship_registry is not None:
+        snap = serialize_relationship_registry(context.relationship_registry)
+        files_written.append(_write_json(relspec_dir / "rules.json", snap, overwrite=True))
+    if context.relationship_contracts is not None:
+        snap = serialize_contract_catalog(context.relationship_contracts)
+        files_written.append(_write_json(relspec_dir / "contracts.json", snap, overwrite=True))
+    if context.compiled_relationship_outputs is not None:
+        snap = serialize_compiled_outputs(context.compiled_relationship_outputs)
+        files_written.append(
+            _write_json(relspec_dir / "compiled_outputs.json", snap, overwrite=True)
+        )
+
+
+def _write_schema_snapshot(
+    schemas_dir: Path,
+    *,
+    name: str,
+    table: pa.Table,
+    files_written: list[str],
+) -> None:
+    doc: JsonDict = {
+        "name": name,
+        "rows": int(table.num_rows),
+        "schema_fingerprint": schema_fingerprint(table.schema),
+        "schema": arrow_schema_to_dict(table.schema),
+    }
+    files_written.append(_write_json(schemas_dir / f"{name}.schema.json", doc, overwrite=True))
+
+
+def _write_schema_group(
+    schemas_dir: Path,
+    *,
+    prefix: str,
+    tables: Mapping[str, pa.Table] | None,
+    files_written: list[str],
+) -> None:
+    if not tables:
+        return
+    for name, table in tables.items():
+        if table is not None:
+            _write_schema_snapshot(
+                schemas_dir,
+                name=f"{prefix}{name}",
+                table=table,
+                files_written=files_written,
+            )
+
+
+def _write_schema_snapshots(
+    bundle_dir: Path, context: RunBundleContext, files_written: list[str]
+) -> None:
+    if not context.include_schemas:
+        return
+    schemas_dir = bundle_dir / "schemas"
+    _ensure_dir(schemas_dir)
+
+    _write_schema_group(
+        schemas_dir,
+        prefix="relspec_input__",
+        tables=context.relspec_input_tables,
+        files_written=files_written,
+    )
+    _write_schema_group(
+        schemas_dir,
+        prefix="relationship_output__",
+        tables=context.relationship_output_tables,
+        files_written=files_written,
+    )
+    _write_schema_group(
+        schemas_dir,
+        prefix="cpg_output__",
+        tables=context.cpg_output_tables,
+        files_written=files_written,
+    )
+
+
 def write_run_bundle(
     *,
-    base_dir: str,
-    run_manifest: Mapping[str, Any],
-    run_config: Mapping[str, Any],
-    # snapshots
-    relationship_registry: Any | None = None,
-    relationship_contracts: Any | None = None,
-    compiled_relationship_outputs: Mapping[str, Any] | None = None,
-    # persisted locations (filesystem mode)
-    relspec_input_locations: Mapping[str, Any] | None = None,
-    # schema snapshots (tables)
-    relspec_input_tables: Mapping[str, pa.Table] | None = None,
-    relationship_output_tables: Mapping[str, pa.Table] | None = None,
-    cpg_output_tables: Mapping[str, pa.Table] | None = None,
-    include_schemas: bool = True,
-    overwrite: bool = True,
-) -> dict[str, Any]:
-    """
-    Writes a “run bundle” directory capturing enough metadata to replay a run later,
-    even if code changes.
+    context: RunBundleContext,
+) -> JsonDict:
+    """Write a run bundle directory capturing enough metadata to replay a run later.
+
+    This works even if code changes.
 
     Directory layout:
       <base_dir>/<bundle_name>/
@@ -371,113 +506,33 @@ def write_run_bundle(
         relspec/contracts.json
         relspec/compiled_outputs.json
         schemas/*.schema.json
+
+    Parameters
+    ----------
+    context:
+        Bundle inputs and snapshot tables.
+
+    Returns
+    -------
+    JsonDict
+        Bundle metadata including paths and files written.
     """
-    _ensure_dir(base_dir)
-    bundle_name = make_run_bundle_name(run_manifest, run_config)
-    bundle_dir = os.path.join(base_dir, bundle_name)
+    base_path = ensure_path(context.base_dir)
+    _ensure_dir(base_path)
+    bundle_name = make_run_bundle_name(context.run_manifest, context.run_config)
+    bundle_dir = base_path / bundle_name
 
-    if overwrite and pathlib.Path(bundle_dir).exists():
-        # remove contents
-        for root, dirs, files in os.walk(bundle_dir, topdown=False):
-            for fn in files:
-                try:
-                    pathlib.Path(os.path.join(root, fn)).unlink()
-                except Exception:
-                    pass
-            for dn in dirs:
-                try:
-                    pathlib.Path(os.path.join(root, dn)).rmdir()
-                except Exception:
-                    pass
-
-    _ensure_dir(bundle_dir)
+    _ensure_bundle_dir(bundle_dir, overwrite=context.overwrite)
 
     files_written: list[str] = []
 
-    # 1) manifest
-    files_written.append(
-        _write_json(os.path.join(bundle_dir, "manifest.json"), dict(run_manifest), overwrite=True)
-    )
-
-    # 2) config (the values that drove this run)
-    files_written.append(
-        _write_json(os.path.join(bundle_dir, "config.json"), dict(run_config), overwrite=True)
-    )
-
-    # 3) repro info (env, pkg versions, git)
-    repo_root = run_config.get("repo_root") if isinstance(run_config, dict) else None
-    repro = collect_repro_info(
-        repo_root=str(repo_root) if repo_root else None, extra={"bundle_name": bundle_name}
-    )
-    files_written.append(_write_json(os.path.join(bundle_dir, "repro.json"), repro, overwrite=True))
-
-    # 4) dataset locations (filesystem mode)
-    if relspec_input_locations is not None:
-        ds_dir = os.path.join(bundle_dir, "datasets")
-        _ensure_dir(ds_dir)
-        locs = serialize_dataset_locations(relspec_input_locations)
-        files_written.append(
-            _write_json(os.path.join(ds_dir, "locations.json"), locs, overwrite=True)
-        )
-
-    # 5) relationship registry + contracts + compiled outputs snapshots
-    relspec_dir = os.path.join(bundle_dir, "relspec")
-    _ensure_dir(relspec_dir)
-
-    if relationship_registry is not None:
-        snap = serialize_relationship_registry(relationship_registry)
-        files_written.append(
-            _write_json(os.path.join(relspec_dir, "rules.json"), snap, overwrite=True)
-        )
-
-    if relationship_contracts is not None:
-        snap = serialize_contract_catalog(relationship_contracts)
-        files_written.append(
-            _write_json(os.path.join(relspec_dir, "contracts.json"), snap, overwrite=True)
-        )
-
-    if compiled_relationship_outputs is not None:
-        snap = serialize_compiled_outputs(compiled_relationship_outputs)
-        files_written.append(
-            _write_json(os.path.join(relspec_dir, "compiled_outputs.json"), snap, overwrite=True)
-        )
-
-    # 6) schema snapshots for tables
-    if include_schemas:
-        schemas_dir = os.path.join(bundle_dir, "schemas")
-        _ensure_dir(schemas_dir)
-
-        def _dump_schema(name: str, t: pa.Table) -> None:
-            doc = {
-                "name": name,
-                "rows": int(t.num_rows),
-                "schema_fingerprint": schema_fingerprint(t.schema),
-                "schema": arrow_schema_to_dict(t.schema),
-            }
-            files_written.append(
-                _write_json(os.path.join(schemas_dir, f"{name}.schema.json"), doc, overwrite=True)
-            )
-
-        # relspec input schemas
-        if relspec_input_tables:
-            for n, t in relspec_input_tables.items():
-                if t is not None:
-                    _dump_schema(f"relspec_input__{n}", t)
-
-        # relationship output schemas
-        if relationship_output_tables:
-            for n, t in relationship_output_tables.items():
-                if t is not None:
-                    _dump_schema(f"relationship_output__{n}", t)
-
-        # cpg output schemas
-        if cpg_output_tables:
-            for n, t in cpg_output_tables.items():
-                if t is not None:
-                    _dump_schema(f"cpg_output__{n}", t)
+    _write_manifest_files(bundle_dir, context, bundle_name=bundle_name, files_written=files_written)
+    _write_dataset_locations(bundle_dir, context, files_written)
+    _write_relspec_snapshots(bundle_dir, context, files_written)
+    _write_schema_snapshots(bundle_dir, context, files_written)
 
     return {
-        "bundle_dir": bundle_dir,
+        "bundle_dir": str(bundle_dir),
         "bundle_name": bundle_name,
         "files_written": files_written,
     }
