@@ -1,20 +1,20 @@
 """Arrow-based DSL for building inference-driven datasets (Acero + compute + contracts)."""
 
-from .contracts import Contract, DedupeSpec, SortKey
-from .dataset_io import compile_to_acero_scan, make_scanner, open_dataset, scan_to_table
-from .expr import E
-from .finalize import FinalizeResult, finalize
-from .joins import JoinSpec, hash_join
-from .kernels import (
+from arrowdsl.contracts import Contract, DedupeSpec, SortKey
+from arrowdsl.dataset_io import compile_to_acero_scan, make_scanner, open_dataset, scan_to_table
+from arrowdsl.expr import E
+from arrowdsl.finalize import FinalizeResult, finalize
+from arrowdsl.joins import JoinSpec, hash_join
+from arrowdsl.kernels import (
     apply_dedupe,
     canonical_sort,
     canonical_sort_if_canonical,
     explode_list_column,
 )
-from .plan import Plan
-from .queryspec import ProjectionSpec, QuerySpec
-from .runner import run_pipeline
-from .runtime import (
+from arrowdsl.plan import Plan
+from arrowdsl.queryspec import ProjectionSpec, QuerySpec
+from arrowdsl.runner import run_pipeline
+from arrowdsl.runtime import (
     DeterminismTier,
     ExecutionContext,
     Ordering,
