@@ -1,13 +1,15 @@
-"""Arrow compute expression helpers for predicates and projections."""
+"""Expression helpers for building Arrow compute predicates."""
 
 from __future__ import annotations
 
 from collections.abc import Sequence
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any
 
-if TYPE_CHECKING:  # pragma: no cover
-    import pyarrow.compute as pc
+import pyarrow as pa
+import pyarrow.compute as pc
+
+type ExpressionLike = str | pc.Expression
+type ScalarValue = bool | int | float | str | bytes | pa.Scalar | None
 
 
 @dataclass(frozen=True)
@@ -16,215 +18,318 @@ class E:
 
     @staticmethod
     def field(name: str) -> pc.Expression:
-        """Return a field expression for a column name.
+        """Return a field reference expression.
+
+        Parameters
+        ----------
+        name:
+            Column name to reference.
 
         Returns
         -------
-        pc.Expression
-            Expression referencing the column.
+        pyarrow.compute.Expression
+            Expression representing the field reference.
         """
-        import pyarrow.compute as pc
-
         return pc.field(name)
 
     @staticmethod
-    def scalar(value: Any) -> pc.Expression:
-        """Return a scalar expression for a literal value.
+    def scalar(value: ScalarValue) -> pc.Expression:
+        """Return a scalar expression.
+
+        Parameters
+        ----------
+        value:
+            Scalar literal value.
 
         Returns
         -------
-        pc.Expression
-            Expression representing the literal.
+        pyarrow.compute.Expression
+            Expression representing the scalar value.
         """
-        import pyarrow.compute as pc
-
         return pc.scalar(value)
 
-    # -------------
-    # Predicates
-    # -------------
-
     @staticmethod
-    def eq(col: str, value: Any) -> pc.Expression:
-        """Return an equality predicate for a column and literal.
+    def eq(col: str, value: ScalarValue) -> pc.Expression:
+        """Return an equality predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Equality predicate expression.
         """
         return E.field(col) == E.scalar(value)
 
     @staticmethod
-    def ne(col: str, value: Any) -> pc.Expression:
-        """Return a not-equal predicate for a column and literal.
+    def ne(col: str, value: ScalarValue) -> pc.Expression:
+        """Return an inequality predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
-            Not-equal predicate expression.
+        pyarrow.compute.Expression
+            Inequality predicate expression.
         """
         return E.field(col) != E.scalar(value)
 
     @staticmethod
-    def lt(col: str, value: Any) -> pc.Expression:
-        """Return a less-than predicate for a column and literal.
+    def lt(col: str, value: ScalarValue) -> pc.Expression:
+        """Return a less-than predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Less-than predicate expression.
         """
         return E.field(col) < E.scalar(value)
 
     @staticmethod
-    def le(col: str, value: Any) -> pc.Expression:
-        """Return a less-than-or-equal predicate for a column and literal.
+    def le(col: str, value: ScalarValue) -> pc.Expression:
+        """Return a less-than-or-equal predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Less-than-or-equal predicate expression.
         """
         return E.field(col) <= E.scalar(value)
 
     @staticmethod
-    def gt(col: str, value: Any) -> pc.Expression:
-        """Return a greater-than predicate for a column and literal.
+    def gt(col: str, value: ScalarValue) -> pc.Expression:
+        """Return a greater-than predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Greater-than predicate expression.
         """
         return E.field(col) > E.scalar(value)
 
     @staticmethod
-    def ge(col: str, value: Any) -> pc.Expression:
-        """Return a greater-than-or-equal predicate for a column and literal.
+    def ge(col: str, value: ScalarValue) -> pc.Expression:
+        """Return a greater-than-or-equal predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        value:
+            Scalar value to compare against.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Greater-than-or-equal predicate expression.
         """
         return E.field(col) >= E.scalar(value)
 
     @staticmethod
-    def between(col: str, lo: Any, hi: Any) -> pc.Expression:
-        """Return a range predicate with inclusive bounds.
+    def between(col: str, lo: ScalarValue, hi: ScalarValue) -> pc.Expression:
+        """Return a closed-interval predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        lo:
+            Lower bound (inclusive).
+        hi:
+            Upper bound (inclusive).
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Range predicate expression.
         """
         return (E.field(col) >= E.scalar(lo)) & (E.field(col) <= E.scalar(hi))
 
     @staticmethod
-    def in_(col: str, values: Sequence[Any]) -> pc.Expression:
-        """Return a membership predicate for a column and values.
+    def in_(col: str, values: Sequence[ScalarValue]) -> pc.Expression:
+        """Return an inclusion predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to compare.
+        values:
+            Candidate values to match.
 
         Returns
         -------
-        pc.Expression
-            Membership predicate expression.
+        pyarrow.compute.Expression
+            Inclusion predicate expression.
         """
         return E.field(col).isin(list(values))
 
     @staticmethod
     def is_null(col: str) -> pc.Expression:
-        """Return a null-check predicate for a column.
+        """Return a null-check predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to check.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Null-check predicate expression.
         """
         return E.field(col).is_null()
 
     @staticmethod
     def is_valid(col: str) -> pc.Expression:
-        """Return a validity-check predicate for a column.
+        """Return a validity-check predicate.
+
+        Parameters
+        ----------
+        col:
+            Column name to check.
 
         Returns
         -------
-        pc.Expression
-            Validity-check predicate expression.
+        pyarrow.compute.Expression
+            Validity predicate expression.
         """
         return E.field(col).is_valid()
 
-    # -------------
-    # Boolean ops
-    # -------------
-
     @staticmethod
     def and_(*exprs: pc.Expression) -> pc.Expression:
-        """Combine expressions with boolean AND.
+        """Combine predicates with logical AND.
+
+        Parameters
+        ----------
+        *exprs:
+            Expressions to combine.
 
         Returns
         -------
-        pc.Expression
-            Conjunction of input expressions.
+        pyarrow.compute.Expression
+            Combined predicate.
+
+        Raises
+        ------
+        ValueError
+            Raised when no expressions are provided.
         """
+        if not exprs:
+            msg = "and_ requires at least one expression."
+            raise ValueError(msg)
         out = exprs[0]
-        for e in exprs[1:]:
-            out = out & e
+        for expr in exprs[1:]:
+            out &= expr
         return out
 
     @staticmethod
     def or_(*exprs: pc.Expression) -> pc.Expression:
-        """Combine expressions with boolean OR.
+        """Combine predicates with logical OR.
+
+        Parameters
+        ----------
+        *exprs:
+            Expressions to combine.
 
         Returns
         -------
-        pc.Expression
-            Disjunction of input expressions.
+        pyarrow.compute.Expression
+            Combined predicate.
+
+        Raises
+        ------
+        ValueError
+            Raised when no expressions are provided.
         """
+        if not exprs:
+            msg = "or_ requires at least one expression."
+            raise ValueError(msg)
         out = exprs[0]
-        for e in exprs[1:]:
-            out = out | e
+        for expr in exprs[1:]:
+            out |= expr
         return out
 
     @staticmethod
     def not_(expr: pc.Expression) -> pc.Expression:
-        """Negate an expression with boolean NOT.
+        """Negate a predicate expression.
+
+        Parameters
+        ----------
+        expr:
+            Expression to negate.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Negated expression.
         """
         return ~expr
 
-    # -------------
-    # Projection helpers
-    # -------------
-
     @staticmethod
-    def cast(expr: str | pc.Expression, target_type: Any, *, safe: bool = True) -> pc.Expression:
-        """Cast an expression to a target Arrow type.
+    def cast(expr: ExpressionLike, target_type: object, *, safe: bool = True) -> pc.Expression:
+        """Cast an expression to a target type.
+
+        Parameters
+        ----------
+        expr:
+            Source expression or column name.
+        target_type:
+            Target Arrow data type.
+        safe:
+            When ``True``, only allow safe casts.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Cast expression.
         """
-        import pyarrow.compute as pc
-
-        e = E.field(expr) if isinstance(expr, str) else expr
-        return pc.cast(e, target_type, safe=safe)
+        expression = E.field(expr) if isinstance(expr, str) else expr
+        return pc.cast(expression, target_type, safe=safe)
 
     @staticmethod
-    def coalesce(*exprs: str | pc.Expression) -> pc.Expression:
-        """Coalesce expressions in order of non-null values.
+    def coalesce(*exprs: ExpressionLike) -> pc.Expression:
+        """Return the first non-null expression in order.
+
+        Parameters
+        ----------
+        *exprs:
+            Expressions or column names to coalesce.
 
         Returns
         -------
-        pc.Expression
+        pyarrow.compute.Expression
             Coalesced expression.
         """
-        import pyarrow.compute as pc
-
-        es = [E.field(e) if isinstance(e, str) else e for e in exprs]
-        return pc.coalesce(*es)
+        expressions = [E.field(expr) if isinstance(expr, str) else expr for expr in exprs]
+        return pc.coalesce(*expressions)
