@@ -1,18 +1,48 @@
-"""Join plan specifications and helpers."""
+"""Join helpers for Acero plans."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass
+from typing import Literal
 
-from .plan import Plan
-from .runtime import Ordering
+from pyarrow import acero
 
-JoinType = str  # e.g. "inner", "left outer", "right outer", "full outer", "left semi", "left anti"
+from arrowdsl.plan import Plan
+from arrowdsl.runtime import Ordering
+
+type JoinType = Literal[
+    "inner",
+    "left outer",
+    "right outer",
+    "full outer",
+    "left semi",
+    "right semi",
+    "left anti",
+    "right anti",
+]
 
 
 @dataclass(frozen=True)
 class JoinSpec:
-    """Specification for hash join inputs and outputs."""
+    """Join specification for hash joins.
+
+    Parameters
+    ----------
+    join_type:
+        Join type string.
+    left_keys:
+        Left-side join keys.
+    right_keys:
+        Right-side join keys.
+    left_output:
+        Output columns from the left side.
+    right_output:
+        Output columns from the right side.
+    output_suffix_for_left:
+        Suffix for left output column name collisions.
+    output_suffix_for_right:
+        Suffix for right output column name collisions.
+    """
 
     join_type: JoinType
     left_keys: tuple[str, ...]
@@ -23,28 +53,45 @@ class JoinSpec:
     output_suffix_for_right: str = ""
 
     def __post_init__(self) -> None:
-        """Validate that left and right key lengths match."""
+        """Validate that left and right key counts match.
+
+        Raises
+        ------
+        ValueError
+            Raised when key lengths differ.
+        """
         if len(self.left_keys) != len(self.right_keys):
-            raise ValueError("left_keys and right_keys must have the same length")
+            msg = "left_keys and right_keys must have the same length."
+            raise ValueError(msg)
 
 
 def hash_join(*, left: Plan, right: Plan, spec: JoinSpec, label: str = "") -> Plan:
-    """Build a hash join declaration within an Acero plan.
+    """Perform a hash join in the plan lane.
+
+    Parameters
+    ----------
+    left:
+        Left-side plan.
+    right:
+        Right-side plan.
+    spec:
+        Join specification.
+    label:
+        Optional plan label.
 
     Returns
     -------
     Plan
-        A plan representing the hash join.
+        Joined plan (unordered output).
 
     Raises
     ------
     TypeError
-        Raised when either plan lacks an Acero declaration.
+        Raised when one or both plans are missing Acero declarations.
     """
-    from pyarrow import acero
-
     if left.decl is None or right.decl is None:
-        raise TypeError("hash_join requires Acero-backed Plans (left.decl and right.decl)")
+        msg = "hash_join requires Acero-backed Plans (missing declarations)."
+        raise TypeError(msg)
 
     decl = acero.Declaration(
         "hashjoin",
@@ -60,5 +107,7 @@ def hash_join(*, left: Plan, right: Plan, spec: JoinSpec, label: str = "") -> Pl
         inputs=[left.decl, right.decl],
     )
     return Plan(
-        decl=decl, label=label or f"{left.label}_join_{right.label}", ordering=Ordering.unordered()
+        decl=decl,
+        label=label or f"{left.label}_join_{right.label}",
+        ordering=Ordering.unordered(),
     )
