@@ -1,3 +1,5 @@
+"""Scan a repository for source files and capture file metadata."""
+
 from __future__ import annotations
 
 import fnmatch
@@ -10,26 +12,37 @@ from pathlib import Path
 
 import pyarrow as pa
 
+from core_types import PathLike, ensure_path
+
 SCHEMA_VERSION = 1
 
 
 def stable_id(prefix: str, *parts: str) -> str:
-    """
-    Stable, deterministic ID builder (content-addressed over string parts).
+    """Build a deterministic string ID.
 
-    Use this for:
-      - file_id
-      - extraction-local ids (cst_def_id, scip_occ_id, etc.)
+    Parameters
+    ----------
+    prefix:
+        ID prefix to namespace the hash.
+    *parts:
+        String parts used for hashing.
+
+    Returns
+    -------
+    str
+        Stable identifier string.
     """
     h = hashlib.sha1()
-    for p in parts:
-        h.update(p.encode("utf-8"))
+    for part in parts:
+        h.update(part.encode("utf-8"))
         h.update(b"\x1f")
     return f"{prefix}:{h.hexdigest()}"
 
 
 @dataclass(frozen=True)
 class RepoScanOptions:
+    """Configure repository scanning behavior."""
+
     repo_id: str | None = None
     include_globs: Sequence[str] = ("**/*.py",)
     exclude_dirs: Sequence[str] = (
@@ -48,19 +61,20 @@ class RepoScanOptions:
     include_bytes: bool = True
     include_text: bool = True
     max_file_bytes: int | None = None
+    max_files: int | None = None
 
 
 REPO_FILES_SCHEMA = pa.schema(
     [
         ("schema_version", pa.int32()),
         ("file_id", pa.string()),
-        ("path", pa.string()),  # repo-relative POSIX
-        ("abs_path", pa.string()),  # absolute path on disk
+        ("path", pa.string()),
+        ("abs_path", pa.string()),
         ("size_bytes", pa.int64()),
-        ("file_sha256", pa.string()),  # sha256 hex of file bytes
+        ("file_sha256", pa.string()),
         ("encoding", pa.string()),
-        ("text", pa.string()),  # decoded best-effort (nullable)
-        ("bytes", pa.binary()),  # raw bytes (nullable if include_bytes=False)
+        ("text", pa.string()),
+        ("bytes", pa.binary()),
     ]
 )
 
@@ -79,32 +93,56 @@ def _sha256_hex(data: bytes) -> str:
 
 
 def _detect_encoding_and_decode(data: bytes) -> tuple[str, str | None]:
-    """
-    Detect encoding using tokenize rules (PEP 263) and decode best-effort.
+    """Detect encoding using tokenize rules and decode best-effort.
+
+    Parameters
+    ----------
+    data:
+        Raw file bytes.
+
+    Returns
+    -------
+    tuple[str, str | None]
+        Detected encoding and decoded text (if available).
     """
     try:
         encoding, _ = tokenize.detect_encoding(io.BytesIO(data).readline)
-    except Exception:
+    except (LookupError, SyntaxError):
         encoding = "utf-8"
     try:
         text = data.decode(encoding, errors="replace")
-        return encoding, text
-    except Exception:
+    except (LookupError, UnicodeError):
         return encoding, None
+    else:
+        return encoding, text
 
 
 def iter_repo_files(repo_root: Path, options: RepoScanOptions) -> Iterator[Path]:
+    """Iterate over repo files that match include/exclude rules.
+
+    Parameters
+    ----------
+    repo_root:
+        Repository root path.
+    options:
+        Scan options.
+
+    Yields
+    ------
+    pathlib.Path
+        Relative paths for matching files.
+    """
     repo_root = repo_root.resolve()
     seen: set[str] = set()
 
     for pat in options.include_globs:
-        for p in repo_root.glob(pat):
-            if p.is_dir():
+        for path in repo_root.glob(pat):
+            if path.is_dir():
                 continue
-            if not options.follow_symlinks and p.is_symlink():
+            if not options.follow_symlinks and path.is_symlink():
                 continue
 
-            rel = p.relative_to(repo_root)
+            rel = path.relative_to(repo_root)
             if _is_excluded_dir(rel, options.exclude_dirs):
                 continue
 
@@ -117,24 +155,36 @@ def iter_repo_files(repo_root: Path, options: RepoScanOptions) -> Iterator[Path]
                 yield rel
 
 
-def scan_repo(repo_root: Path, options: RepoScanOptions | None = None) -> pa.Table:
-    """
-    Scans repo for Python files and returns a "repo_files" Arrow table.
+def scan_repo(repo_root: PathLike, options: RepoScanOptions | None = None) -> pa.Table:
+    """Scan the repo for Python files and return a repo_files table.
 
-    Output contract is used by all downstream extractors.
+    Parameters
+    ----------
+    repo_root:
+        Repository root path.
+    options:
+        Scan options.
+
+    Returns
+    -------
+    pyarrow.Table
+        Table of repo file metadata.
     """
     options = options or RepoScanOptions()
-    repo_root = repo_root.resolve()
+    repo_root_path = ensure_path(repo_root).resolve()
 
-    rows: list[dict] = []
-    for rel in sorted(iter_repo_files(repo_root, options), key=lambda p: p.as_posix()):
-        abs_path = (repo_root / rel).resolve()
+    max_files = options.max_files
+    if max_files is not None and max_files <= 0:
+        return pa.Table.from_pylist([], schema=REPO_FILES_SCHEMA)
+
+    rows: list[dict[str, object]] = []
+    for rel in sorted(iter_repo_files(repo_root_path, options), key=lambda p: p.as_posix()):
+        abs_path = (repo_root_path / rel).resolve()
         rel_posix = rel.as_posix()
 
         try:
             data = abs_path.read_bytes()
-        except Exception:
-            # unreadable file -> skip; you can also choose to emit an error table later
+        except OSError:
             continue
 
         if options.max_file_bytes is not None and len(data) > options.max_file_bytes:
@@ -148,17 +198,20 @@ def scan_repo(repo_root: Path, options: RepoScanOptions | None = None) -> pa.Tab
         if options.include_text:
             encoding, text = _detect_encoding_and_decode(data)
 
-        row = {
-            "schema_version": SCHEMA_VERSION,
-            "file_id": file_id,
-            "path": rel_posix,
-            "abs_path": str(abs_path),
-            "size_bytes": len(data),
-            "file_sha256": file_sha256,
-            "encoding": encoding,
-            "text": text,
-            "bytes": data if options.include_bytes else None,
-        }
-        rows.append(row)
+        rows.append(
+            {
+                "schema_version": SCHEMA_VERSION,
+                "file_id": file_id,
+                "path": rel_posix,
+                "abs_path": str(abs_path),
+                "size_bytes": len(data),
+                "file_sha256": file_sha256,
+                "encoding": encoding,
+                "text": text,
+                "bytes": data if options.include_bytes else None,
+            }
+        )
+        if max_files is not None and len(rows) >= max_files:
+            break
 
     return pa.Table.from_pylist(rows, schema=REPO_FILES_SCHEMA)
