@@ -1,6 +1,8 @@
+"""Hamilton pipeline entry points for CodeAnatomy."""
+
 from __future__ import annotations
 
-from .driver_factory import (
+from hamilton_pipeline.driver_factory import (
     DriverFactory,
     build_driver,
     config_fingerprint,
