@@ -1,17 +1,40 @@
+"""Hamilton normalization stage functions."""
+
 from __future__ import annotations
 
 import pyarrow as pa
-
 from hamilton.function_modifiers import cache, extract_fields, tag
 
-from ...arrowdsl.runtime import ExecutionContext
-from ...normalize.spans import (
+from arrowdsl.kernels import explode_list_column
+from arrowdsl.runtime import ExecutionContext
+from extract.repo_scan import stable_id
+from normalize.spans import (
     RepoTextIndex,
     add_scip_occurrence_byte_spans,
     build_repo_text_index,
     normalize_cst_defs_spans,
     normalize_cst_imports_spans,
 )
+
+
+def _collect_string_lists(table: pa.Table | None, column: str) -> list[str]:
+    """Flatten a list-valued column into a list of strings.
+
+    Returns
+    -------
+    list[str]
+        Flattened string values.
+    """
+    if table is None or column not in table.column_names:
+        return []
+    out: list[str] = []
+    for items in table[column].to_pylist():
+        if not isinstance(items, list):
+            continue
+        if not items:
+            continue
+        out.extend(str(item) for item in items if item)
+    return out
 
 
 @cache()
@@ -21,32 +44,20 @@ def dim_qualified_names(
     cst_defs: pa.Table,
     ctx: ExecutionContext,
 ) -> pa.Table:
-    """
-    Build a dimension table of qualified names encountered in CST extraction.
+    """Build a dimension table of qualified names from CST extraction.
 
     Expected output columns:
       - qname_id (stable)
       - qname (string)
+
+    Returns
+    -------
+    pa.Table
+        Qualified name dimension table.
     """
-    qnames: list[str] = []
-
-    # From callsites: callee_qnames may be a list of strings per row
-    if cst_callsites is not None and "callee_qnames" in cst_callsites.column_names:
-        for lst in cst_callsites["callee_qnames"].to_pylist():
-            if not lst:
-                continue
-            for s in lst:
-                if s:
-                    qnames.append(str(s))
-
-    # From defs: qnames may be list[str]
-    if cst_defs is not None and "qnames" in cst_defs.column_names:
-        for lst in cst_defs["qnames"].to_pylist():
-            if not lst:
-                continue
-            for s in lst:
-                if s:
-                    qnames.append(str(s))
+    _ = ctx
+    qnames = _collect_string_lists(cst_callsites, "callee_qnames")
+    qnames.extend(_collect_string_lists(cst_defs, "qnames"))
 
     if not qnames:
         return pa.Table.from_pylist(
@@ -69,8 +80,7 @@ def callsite_qname_candidates(
     cst_callsites: pa.Table,
     ctx: ExecutionContext,
 ) -> pa.Table:
-    """
-    Explode cst_callsites.callee_qnames into a row-per-candidate table.
+    """Explode callsite qualified names into a row-per-candidate table.
 
     Output columns (minimum):
       - call_id
@@ -79,7 +89,13 @@ def callsite_qname_candidates(
       - call_bstart
       - call_bend
       - qname_source (optional)
+
+    Returns
+    -------
+    pa.Table
+        Table of callsite qualified name candidates.
     """
+    _ = ctx
     if (
         cst_callsites is None
         or cst_callsites.num_rows == 0
@@ -146,27 +162,39 @@ def callsite_qname_candidates(
 @cache()
 @tag(layer="normalize", artifact="repo_text_index", kind="object")
 def repo_text_index(repo_root: str, repo_files: pa.Table, ctx: ExecutionContext) -> RepoTextIndex:
-    """
-    Builds a repo text index used to convert (line,col) → byte offsets.
+    """Build a repo text index for line/column to byte offsets.
+
+    Returns
+    -------
+    RepoTextIndex
+        Repository text index for span conversion.
     """
     return build_repo_text_index(repo_root=repo_root, repo_files=repo_files, ctx=ctx)
 
 
 @cache()
 @extract_fields(
-    scip_occurrences_norm=pa.Table,
-    scip_span_errors=pa.Table,
+    {
+        "scip_occurrences_norm": pa.Table,
+        "scip_span_errors": pa.Table,
+    }
 )
 @tag(layer="normalize", artifact="scip_occurrences_norm_bundle", kind="bundle")
 def scip_occurrences_norm_bundle(
+    scip_documents: pa.Table,
     scip_occurrences: pa.Table,
     repo_text_index: RepoTextIndex,
     ctx: ExecutionContext,
 ) -> dict[str, pa.Table]:
-    """
-    Converts SCIP occurrence (line/col spans) into byte offsets bstart/bend.
+    """Convert SCIP occurrences into byte offsets.
+
+    Returns
+    -------
+    dict[str, pa.Table]
+        Bundle with normalized occurrences and span errors.
     """
     occ, errs = add_scip_occurrence_byte_spans(
+        scip_documents=scip_documents,
         scip_occurrences=scip_occurrences,
         repo_text_index=repo_text_index,
         ctx=ctx,
@@ -177,16 +205,26 @@ def scip_occurrences_norm_bundle(
 @cache()
 @tag(layer="normalize", artifact="cst_imports_norm", kind="table")
 def cst_imports_norm(cst_imports: pa.Table, ctx: ExecutionContext) -> pa.Table:
+    """Normalize CST import spans into bstart/bend.
+
+    Returns
+    -------
+    pa.Table
+        Normalized CST imports table.
     """
-    Normalizes CST import spans into bstart/bend (alias span preferred).
-    """
-    return normalize_cst_imports_spans(cst_imports=cst_imports, ctx=ctx)
+    _ = ctx
+    return normalize_cst_imports_spans(py_cst_imports=cst_imports)
 
 
 @cache()
 @tag(layer="normalize", artifact="cst_defs_norm", kind="table")
 def cst_defs_norm(cst_defs: pa.Table, ctx: ExecutionContext) -> pa.Table:
+    """Normalize CST def spans into bstart/bend.
+
+    Returns
+    -------
+    pa.Table
+        Normalized CST definitions table.
     """
-    Normalizes CST def spans into bstart/bend.
-    """
-    return normalize_cst_defs_spans(cst_defs=cst_defs, ctx=ctx)
+    _ = ctx
+    return normalize_cst_defs_spans(py_cst_defs=cst_defs)
