@@ -1,33 +1,14 @@
-from __future__ import annotations
+"""Normalization helpers for extracted tables."""
 
-"""
-normalize/
-
-This package is where extraction tables become join-ready for relationship rules and Acero plans.
-
-Core responsibilities:
-  1) Spans: normalize all source ranges to *byte offsets* (bstart/bend) in UTF-8 bytes.
-     - AST: lineno/col_offset -> byte spans
-     - SCIP: (line,char) ranges -> byte spans using Document.position_encoding semantics
-     - CST: already byte spans, but we add canonical aliases (bstart/bend) for consistent joins
-  2) Schema inference: unify schemas across partially-populated tables and align tables permissively.
-  3) Stable IDs: deterministic hashing for ids/keys/edges, so downstream graph building is reproducible.
-"""
-
-from .ids import (
-    add_span_id_column,
-    span_id,
-    stable_id,
-    stable_int64,
-)
-from .schema_infer import (
+from normalize.ids import add_span_id_column, span_id, stable_id, stable_int64
+from normalize.schema_infer import (
     SchemaInferOptions,
     align_table_to_schema,
     align_tables_to_unified_schema,
     infer_schema_from_tables,
     unify_schemas,
 )
-from .spans import (
+from normalize.spans import (
     FileTextIndex,
     RepoTextIndex,
     add_ast_byte_spans,
@@ -39,24 +20,21 @@ from .spans import (
 )
 
 __all__ = [
-    # spans
     "FileTextIndex",
     "RepoTextIndex",
-    "build_repo_text_index",
+    "SchemaInferOptions",
     "add_ast_byte_spans",
     "add_scip_occurrence_byte_spans",
-    "normalize_cst_callsites_spans",
-    "normalize_cst_imports_spans",
-    "normalize_cst_defs_spans",
-    # schema inference
-    "SchemaInferOptions",
-    "unify_schemas",
-    "infer_schema_from_tables",
+    "add_span_id_column",
     "align_table_to_schema",
     "align_tables_to_unified_schema",
-    # ids
+    "build_repo_text_index",
+    "infer_schema_from_tables",
+    "normalize_cst_callsites_spans",
+    "normalize_cst_defs_spans",
+    "normalize_cst_imports_spans",
+    "span_id",
     "stable_id",
     "stable_int64",
-    "span_id",
-    "add_span_id_column",
+    "unify_schemas",
 ]
