@@ -1,15 +1,22 @@
+"""Stable identifier helpers for normalization stages."""
+
 from __future__ import annotations
 
 import hashlib
+from dataclasses import dataclass
 
 import pyarrow as pa
 
 
 def stable_id(prefix: str, *parts: str) -> str:
-    """
-    Deterministic string ID.
+    """Build a deterministic string ID.
 
     Must remain stable across runs/machines and should be used for any persisted identifiers.
+
+    Returns
+    -------
+    str
+        Stable identifier with the requested prefix.
     """
     h = hashlib.sha1()
     for p in parts:
@@ -19,10 +26,14 @@ def stable_id(prefix: str, *parts: str) -> str:
 
 
 def stable_int64(*parts: str) -> int:
-    """
-    Deterministic signed int64 derived from parts.
+    """Build a deterministic signed int64 from parts.
 
     Useful as compact row ids for Arrow joins if you prefer numeric keys.
+
+    Returns
+    -------
+    int
+        Deterministic signed 64-bit integer.
     """
     h = hashlib.sha1()
     for p in parts:
@@ -33,36 +44,66 @@ def stable_int64(*parts: str) -> int:
     return int(v)
 
 
+def _row_value_int(value: object | None) -> int | None:
+    if isinstance(value, bool):
+        return None
+    if isinstance(value, int):
+        return value
+    if isinstance(value, float) and value.is_integer():
+        return int(value)
+    if isinstance(value, str) and value.isdigit():
+        return int(value)
+    return None
+
+
 def span_id(path: str, bstart: int, bend: int, kind: str | None = None) -> str:
-    """
-    Stable ID for a source span, optionally keyed by kind.
+    """Build a stable ID for a source span.
 
     This is *not required* as a join key (path+bstart+bend is enough),
     but it is useful for:
       - CPG node ids anchored to code
       - reproducible edge ids
       - cache keys
+
+    Returns
+    -------
+    str
+        Stable span identifier.
     """
     if kind:
         return stable_id("span", kind, path, str(bstart), str(bend))
     return stable_id("span", path, str(bstart), str(bend))
 
 
-def add_span_id_column(
-    table: pa.Table,
-    *,
-    path_col: str = "path",
-    bstart_col: str = "bstart",
-    bend_col: str = "bend",
-    kind: str | None = None,
-    out_col: str = "span_id",
-) -> pa.Table:
-    """
-    Adds a span_id column computed row-by-row.
+@dataclass(frozen=True)
+class SpanIdSpec:
+    """Specification for span_id column generation."""
+
+    path_col: str = "path"
+    bstart_col: str = "bstart"
+    bend_col: str = "bend"
+    kind: str | None = None
+    out_col: str = "span_id"
+
+
+def add_span_id_column(table: pa.Table, spec: SpanIdSpec | None = None) -> pa.Table:
+    """Add a span_id column computed row-by-row.
 
     Note: this is Python-loop based; for very large tables you may want
     to compute ids later or in batches.
+
+    Returns
+    -------
+    pa.Table
+        Table with the span_id column appended.
     """
+    spec = spec or SpanIdSpec()
+    path_col = spec.path_col
+    bstart_col = spec.bstart_col
+    bend_col = spec.bend_col
+    kind = spec.kind
+    out_col = spec.out_col
+
     if out_col in table.column_names:
         return table
 
@@ -79,10 +120,12 @@ def add_span_id_column(
     )
 
     out: list[str | None] = []
-    for p, bs, be in zip(path_arr, bs_arr, be_arr):
-        if p is None or bs is None or be is None:
+    for p, bs, be in zip(path_arr, bs_arr, be_arr, strict=True):
+        bs_int = _row_value_int(bs)
+        be_int = _row_value_int(be)
+        if p is None or bs_int is None or be_int is None:
             out.append(None)
             continue
-        out.append(span_id(str(p), int(bs), int(be), kind=kind))
+        out.append(span_id(str(p), bs_int, be_int, kind=kind))
 
     return table.append_column(out_col, pa.array(out, type=pa.string()))
